requirements.txt
flask
google-cloud-bigquery
pandas
python-dotenv
pyarrow
